scripts/requirements.txt
google_api_python_client==2.129.0
google_auth_oauthlib==1.2.0
numpy==1.26.4
openai==1.30.3
pandas==2.2.2
protobuf==5.27.0
python_dateutil==2.8.2
pytz==2024.1
Requests==2.32.2
scikit_learn==1.4.1.post1
